feat(filter): Separate check boxes and picker view for reusing
